test(flaky): Add timeouts to gossip smoke test (#1364)
## Description

This test is seen to be flaky on CI and just spins forever until CI
times out and gives up.  By adding timeouts to the test itself we
should see the test fail instead of spin forever.  And as a failed
test we should get logs from CI.  Logs are probably needed to figure
out why this is flaky.

<!-- A summary of what this pull request achieves and a rough list of
changes. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.