fix(sync): fix `PeerData` encoding, neighbor events, better & predictable tests (#1513)
## Description

* **fix(gossip): properly encode peer data.** #1506 introduced a bug:
The `PeerData` was encoded from the `PeerAddr` (including the peer id)
but decoded to `AddrInfo` (without the peer id). So it failed, and
dialing peers failed. It only did not matter much because most tests use
tickets separately, so do not rely on the `PeerData` gossip.
* **feat: expose neighbor events** through document subscriptions. for
now used to write better and less flakey tests. also useful for stats
like usecases, and potentially others.
* **tests: improve sync tests** and make `sync_full_basic` not flakey
anymore (hopefully). the main change is that for some events, we don't
care about the exact order in tests anymore, because the exact order is
too unpredictable timing-wise for things that happen concurrently.
instead they are matched in chunks.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.

Replaces #1512