refactor(iroh-net): rename PeerAddr to NodeAddr, introduce NodeId alias (#1765)
## Description
First step on dealing with the whole peer_id vs node_id. Changes
include:
- renames `PeerAddr` to `NodeAddr` and it's inner fields from `peer_id`
to `node_id`
- introduces the `NodeId` type alias as stated in #1388 
- moves every field and parameter defined as `peer_id: PublicKey` to
`node_id: NodeId`. As far as I can tell everywhere touched is not doing
any crypto stuff.

## Notes & open questions

The one place where this is not changed is the tls code as it does deal
with crypto ops. Later refactors should update the code to use `key:
PublicKey` instead of `peer_id: PublicKey` for example.

In the future it might be worth to actually rename `PublicKey` to
`NodeId` or creating a newtype `NodeId` that can only be obtained from a
`PublicKey` but has no crypto ops available. Not doing any of that since
I know it's controversial.

Also note that this is intentionally kept partial. Reasons for this:
- the state already has a mixed peer_id/node_id state, so we are not
passing from a consistent state into an inconsistent one, but from a
less desirable and inconsistent one to a more desirable, still
inconsistent one
- keeping the PR short to facilitate a meaningful review

## Change checklist

- [x] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.