refactor(iroh-net)!: Rename Endpoint::local_endpoints to direct_addresses (#2369)
## Description

This renames Endpoint::local_endpoints to Endpoint::direct_addresses. As
a consequence it renames a lot of other things, e.g. config::Endpoint
type becomes magicsock::DirectAddr. This is hopefully the last ambiguous
use of "endpoint". The name "direct addresses" seemed to be the
consensus on a discord bikeshedding thread.

The entire config module is removed and the types are moved into
magicsock instead. These types did not have anything to do with
configuration and this was also a source of confusion. Because the
visibility of these types changed some dead code was removed as well.

## Breaking Changes

- iroh_net::Endpoint::local_endpoints ->
iroh_net::Endpoint::direct_addresses
- iroh_net::endpoint::LocalEndpointStream ->
iroh_net::endpoint::DirectAddrStream
- iroh_gossip::net::Gossip::update_endpoints ->
iroh_gossip::net::Gossip::update_direct_addresses
- iroh_net::config is removed.
- iroh_net::config::Endpoint -> iroh_net::magicsock::DirectAddr
- iroh_net::config::EndpointType -> iroh_net::magicsock::DirectAddrType
- iroh_net::config::NetInfo -> removed
- iroh_net::config::LinkInfo -> removed

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- ~~[ ] Tests if relevant.~~
- [x] All breaking changes documented.