fix(iroh-gossip): gossip dispatcher: reliable events on join, allow dropping sinks or streams (#2482)
## Description

This has 2 fixes to the gossip dispatcher:

* When a topic is joined, it can currently happen that the first few
events on the newly joined topic are missed, because the events can
start to arrive at the dispatcher's global subscription before the
dispatcher's join task finished and changed the protocol state. If this
happens, we now change the topic state to live before processing the
incoming event. This should make the existing smoke test non-flaky.

* Currently a topic subscription is quitted when either the command
stream (from client to gossip) or the event stream (from gossip to
client) is dropped. Now we only quit the subscription if both the
command and the event channels are dropped. This should fix the test in
iroh-ffi.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Tests if relevant.