fix: fix reconnects on the proto level
This changes the hyparview logic to send a `Neighbor` message in reply
to a `Join` even if the joining node is already in the active view.
Previously, no reply was sent in this case. With this change, nodes can
reconnect fine, whereas before, they could not reconnect unless their
bootstrap node had them removed from the active view after timeout.