refactor(hyparview): Improve disconnect handling
This improves the abstractions around when Disconnect messages are sent
and when peers are disconnected. Through a new RemovalReason enum, we
can handle this much cleaner. We also store in the local state if we
intended to disconnect from a peer, to not remove that peer from our
passive view once the peer is actually disconnected on the network layer.

Part of https://github.com/n0-computer/iroh-gossip/pull/53