tests(net): Add a net test with many nodes (#65)
This adds a large integration test that bootstraps a swarm with many nodes
and lets each node broadcast some messages.

The test runs with 100 nodes and 2 messages per node.

On my machine I ran it a couple of times with 1000 nodes in release mode.
It completes fine, but takes a while in debug mode,
thus limited to 100 nodes for now to not overwhelm CI.