fix(plumtree): Ensure eager relation is symmetrical
Small fix to plumtree implementation: Make sure that when we send out a
Graft message, we also mark the peer eager in our state. Same the other
way round: When we send a Prune message, mark the peer as lazy in our
state.

This commit is part of https://github.com/n0-computer/iroh-gossip/pull/53