fix(hyparview)!: Only add peers to active view after receiving neighbor messages (#56)
This is a change in behavior of hyparview. Currently, and according to the paper,
peers are added to a node's active view directly after receiving a ForwardJoin
for the peer. This means that a peer is added without knowing if the peer is
reachable or online. This PR changes this behavior so that on ForwardJoin,
we send out a Neighbor message, but wait until we receive the Neighbor reply
to add the peer to our active view. The PR also forward failed dial events to
the protocol state, to clean up state.

A consequence is that the joining phase after a ForwardJoin takes one roundtrip
to complete, whereas before we would start sending messages to our new peer right
after the connection is established. I think this delay is worth the gained assurance
that NeighborUp events are only emitted for peers that we can actually talk to.

This substantially improves the protocol according to simulations, as we
don't collect stale peers in our active view.

While this is not a wire- or API-breaking change, this changes the
behavior as to which messages are sent out when. While I *think* that
nodes before and after this change will work well together, we did not
verify or test this yet.

This commit is part of https://github.com/n0-computer/iroh-gossip/pull/53