refactor: Use discovery service instead of `Endpoint::add_node_addr` (#108)
## Description

iroh-gossip includes a node's addressing information in all Join and ForwardJoin messages (in the peer_data field). This is used by receives of the messages to be able to dial the sender directly, even when not using external discovery mechanisms.

We are using `Endpoint::add_node_addr_with_source` to add the received addr info to the endpoint. In iroh 0.93, `Endpoint::add_node_addr` was removed, and `add_node_addr_with_source` *should* have been removed as well, but was overlooked, and thus iroh-gossip 0.93 only works like previously right now because of an oversight in https://github.com/n0-computer/iroh/pull/3338. We will remove `add_node_addr_with_source` in 0.94 and maybe already deprecate sooner.

This PR changes iroh-gossip to not use `Endpoint::add_node_addr_with_source` anymore and instead get the received addressing info to the endpoint through a discoery service.

The discovery service is fully internal atm: It is not exposed, only added to the endpoint and used from within the crate.

I first used a `StaticProvider` but that would keep the records forever, which is a memory leak for a long-running busy gossip node.  So I switched to an impl that evicts entries after some time.

Right now I hard-coded the retention to 5 minutes, evicting expired entries every 30s. Should this be configurable? Maybe yes, I think.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them, including how to migrate older code. -->

## Notes & open questions

Should the retention settings be configurable? Could quite easily add it to the builder.

Previously, this "retention" was fully outside the control of either iroh-gossip or the user, as it was handled by iroh's magic socket internal pruning of unused entries from the node map. Would need to look up what intervals that are.



## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text), if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.