fix: DisconnectPeer out-event on closed connection
So far OutEvent::DisconnectPeer was only moved into the outbox queue
when HyParView triggered (random) re-shuffle, moving active peers into
the passive view when a new peer joined, or if an explicit disconnect
message was received.

Unfortunately this didn't cover (unexpected) closing of connections, for
example when a timeout occurred.

This fix pushes an OutEvent::DisconnectPeer into the outbox queue for
exactly that case when a connection got closed (for unknown reasons).