fix: Logic to remove peer when no topics are left
`HashSet::remove` returns false if the topic in the set was not present
anymore. The logical AND (&&) for checking if the topic list is empty
didn't catch that case and would result in empty = false, even though
the list _is_ actually empty.

Changing it to an OR (||) fixes this issue and properly cleans up
unused connections when processing `DisconnectedPeer` out-events.